Created a CancelButton class
